Fixed duplication
