+++ src/test/java/io/percy/appium/checkstyles.xml
@@ -165,7 +165,7 @@
     <!-- See https://checkstyle.org/config_design.html -->
     <module name="FinalClass"/>
     <module name="InterfaceIsType"/>
-    <module name="VisibilityModifier"/>
+    <!-- <module name="VisibilityModifier"/> -->
 
     <!-- Miscellaneous other checks.                   -->
     <!-- See https://checkstyle.org/config_misc.html -->
